Pb with 'protected' of  circlular_list_head<T>
Moving corresponding head list specific methods to the circular_list_head<T> class was not possible by keeping the "protected" keyword for the circular_list<T>'s attributes.  This is because lot of actions done by these methods imply manipulating these attributes through a pointer to an instance of the base type instead of through the 'this' only. In that case the 'protected' access doesn't allow to access to these attributes. This is well confirmed by the c++ reference's description of the "protected' keyword.